refactor(자료형 변경) : 레이싱을 진행할 수 있는 숫자 범위 제한
결과 출력 시 메모리 문제가 있기 때문에 최대 레이싱 진행 횟수를 Integer형의 Max값으로 제한

+++ src/main/java/racingcar/model/car/Car.java
@@ -5,5 +5,5 @@ public interface Car {
 
     String getName();
 
-    Long getProgress();
+    Integer getProgress();
 }
